requirements.txt
streamlit
langchain
langchain-community
langgraph

chromadb
sentence-transformers
pypdf

torch
torchvision
transformers
Pillow